Moving the final move back to complete
