+++ lib/complete.js
@@ -10,20 +10,44 @@
 'use strict';
 
 var exec = require('child_process').exec;
+var async = require('async');
+var mv = require('mv');
 
 module.exports = function (state, callback) {
     var complete = state.data.complete;
 
-    if (complete !== undefined) {
-        exec(complete, function (err) {
-            if (err) {
-                callback(err);
-                return;
+    async.series([
+        // Move all the files to the relative path "{{ name }}"
+        function (next) {
+            mv(state.template.tmp, state.data.name, function (err) {
+                if (err) {
+                    next(err);
+                    return;
+                }
+                next(null);
+            });
+        },
+        // Call the shell command specified as `complete`
+        function (next) {
+            if (complete !== undefined) {
+                exec(complete, function (err) {
+                    if (err) {
+                        next(err);
+                        return;
+                    }
+                    next(null);
+                });
+            } else {
+                next(null);
             }
-            callback(null, state);
-        });
-    } else {
+        }
+    ], function (err) {
+        if (err) {
+            callback(err);
+            return;
+        }
         callback(null, state);
-    }
+    });
+
 };
