+++ lib/move.js
@@ -14,58 +14,37 @@ var mv = require('mv');
 var async = require('async');
 var fs = require('fs');
 var mkdirp = require('mkdirp');
-var mv = require('mv');
+var async = require('async');
 
 module.exports = function (state, callback) {
     var move = state.data.move;
 
-    async.series([
-        // Move around all the items specified in the schema
-        function (next) {
-            if (move === undefined || move.length === 0) {
-                next(null);
-                return;
-            }
+    if (move === undefined || move.length === 0) {
+        callback(null, state);
+        return;
+    }
 
-            async.parallel(move.map(function (item) {
-                var orig = state.template.tmp + '/' + Object.keys(item)[0],
-                    dest = state.template.tmp + '/' + item[Object.keys(item)[0]],
-                    destPathOnly;
+    async.parallel(move.map(function (item) {
+        var orig = state.template.tmp + '/' + Object.keys(item)[0],
+            dest = state.template.tmp + '/' + item[Object.keys(item)[0]],
+            destPathOnly;
 
-                if (fs.existsSync(orig)) {
-                    destPathOnly = dest.substring(0, dest.lastIndexOf('/'));
-                } else {
-                    destPathOnly = dest;
-                }
-
-                return function (done) {
-                    mkdirp(destPathOnly, function (err) {
-                        if (err) {
-                            done(err);
-                            return;
-                        }
-                        mv(orig, dest, done);
-                    });
-                };
-            }), function (err) {
-                if (err) {
-                    next(err);
-                    return;
-                }
-                next(null);
-            });
-        },
-        // Move all the files to the relative path "{{ name }}"
-        function (next) {
-            mv(state.template.tmp, state.data.name, function (err) {
-                if (err) {
-                    next(err);
-                    return;
-                }
-                next(null);
-            });
+        if (fs.existsSync(orig)) {
+            destPathOnly = dest.substring(0, dest.lastIndexOf('/'));
+        } else {
+            destPathOnly = dest;
         }
-    ], function (err) {
+
+        return function (done) {
+            mkdirp(destPathOnly, function (err) {
+                if (err) {
+                    done(err);
+                    return;
+                }
+                mv(orig, dest, done);
+            });
+        };
+    }), function (err) {
         if (err) {
             callback(err);
             return;
